feat: make sure rewrite rule is applied

+++ tests/cases/standalone/common/function/json/json_get.sql
@@ -73,13 +73,13 @@ SELECT json_get_int(json_get_object(j, '[0]'), 'a.i') FROM jsons;
 
 SELECT json_get_int(json_get_object(j, '[9]'), 'a.i') FROM jsons;
 
-SELECT json_get(j, '[0]')::int FROM jsons;
+SELECT json_get(j, '[0]')::int8 FROM jsons;
 
-SELECT json_get(j, '[0]')::float FROM jsons;
+SELECT json_get(j, '[0]')::double FROM jsons;
 
-SELECT json_get(j, '[1]')::int FROM jsons;
+SELECT json_get(j, '[1]')::int8 FROM jsons;
 
-SELECT json_get(j, '[1]')::float FROM jsons;
+SELECT json_get(j, '[1]')::double FROM jsons;
 
 SELECT json_get(j, '[2]')::bool FROM jsons;
